testing 3

+++ src/util/util.go
@@ -147,8 +147,8 @@ func GetUserType(uid, clusterid int64) (string, error) {
 	if jsonp.Path("data.group_id").Data() == nil {
 		return "", errors.New("oweruser id")
 	}
-	log.Debug("-----------+++", jsonp.Path("data.group_id").Data().(int64))
-	return fmt.Sprintf("%d", jsonp.Path("data.group_id").Data().(int64)), nil
+	log.Debug("-----------+++", jsonp.Path("data.group_id").Data().(float64))
+	return fmt.Sprintf("%d", jsonp.Path("data.group_id").Data().(float64)), nil
 }
 
 func CronTokenBuilder(uid string) string {
